% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_data.R
\name{complete_data}
\alias{complete_data}
\title{Complete Data}
\usage{
complete_data(data)
}
\arguments{
\item{data}{data.frame contenant les colonnes "time" et "value"}
}
\value{
Un data.frame avec les dates manquantes ajoutées et les valeurs associées définies comme NA
}
\description{
Ajouter les dates manquantes à un ensemble de données en détectant la fréquence des données
}
\examples{
complete_data(my_data)

}
