% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate model}
\usage{
evaluate_model(model, mydata, ntree = NULL)
}
\arguments{
\item{model}{model that you want to train from caret package (see names(getModelInfo()))}

\item{mydata}{data.frame containing value and time columns}
}
\value{
RMSE
}
\description{
La fonction est utilisée pour évaluer un modèle donné en utilisant une mesure de qualité, RMSE.
}
\examples{
evaluate_model("rf", data)

}
