% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_values.R
\name{impute_values}
\alias{impute_values}
\title{Impute values}
\usage{
impute_values(model, mydata, ntree = NULL)
}
\arguments{
\item{model}{a string representing the model that you want to train from the caret package (see names(getModelInfo()))}

\item{mydata}{a data.frame containing the columns 'value' and 'time'}

\item{ntree}{an optional integer argument for the number of trees to use in a random forest model}
}
\description{
Interpolates missing values. Can either use machine learning models or interpolation methods (spline and approximation) to impute the missing values.
}
\examples{
impute_values("rf", data)

}
